+++ CHANGELOG.md
@@ -9,6 +9,10 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 -   Add support for Edge in Windows (`--browser=edge`).
 
+-   Add support for remote WebDriver with e.g.
+    `--browser=chrome@http://<remote-selenium-server>`. See `README` for more
+    details.
+
 -   Fix bug where no browser other than Chrome could be launched.
 
 -   Fix bug where process did not exit on most exceptions.