+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+<!-- ## Unreleased -->
+<!-- Add new, unreleased changes here. -->
+
+## [0.4.6] 2019-06-12
 
 -   Add support for Edge in Windows (`--browser=edge`).
 
@@ -20,8 +23,6 @@ project adheres to [Semantic Versioning](http://semver.org/).
 -   Fix bug where no browser other than Chrome could be launched.
 
 -   Fix bug where process did not exit on most exceptions.
-
-<!-- Add new, unreleased changes here. -->
 
 ## [0.4.5] 2019-06-10